// src/generated/stream_list.ts
export enum LiveChatMessageType {
  INVALID_TYPE = 0,
  TEXT_MESSAGE_EVENT = 1,
  TOMBSTONE = 2,
  CHAT_ENDED_EVENT = 4,
  SPONSOR_ONLY_MODE_STARTED_EVENT = 5,
  SPONSOR_ONLY_MODE_ENDED_EVENT = 6,
  NEW_SPONSOR_EVENT = 7,
  USER_BANNED_EVENT = 10,
  SUPER_CHAT_EVENT = 15,
  SUPER_STICKER_EVENT = 16,
  MEMBER_MILESTONE_CHAT_EVENT = 17,
  MEMBERSHIP_GIFTING_EVENT = 18,
  GIFT_MEMBERSHIP_RECEIVED_EVENT = 19,
  POLL_EVENT = 20,
  GIFT_EVENT = 21,
  UNRECOGNIZED = -1,
}
// ...
export interface LiveChatMessageListResponse {
  offlineAt?: string | undefined;
  nextPageToken?: string | undefined;
  items: LiveChatMessage[];
}

export interface LiveChatMessage {
  id?: string | undefined;
  snippet?: LiveChatMessageSnippet | undefined;
  authorDetails?: LiveChatMessageAuthorDetails | undefined;
}

export interface LiveChatMessageAuthorDetails {
  channelId?: string | undefined;
  displayName?: string | undefined;
  isVerified?: boolean | undefined;
  isChatOwner?: boolean | undefined;
  isChatSponsor?: boolean | undefined;
  isChatModerator?: boolean | undefined;
}

export interface LiveChatMessageSnippet {
  type?: LiveChatMessageType | undefined;
  publishedAt?: string | undefined;
  displayMessage?: string | undefined;
  hasDisplayContent?: boolean | undefined;
}
// ...
export const LiveChatMessageListResponse: MessageFns<LiveChatMessageListResponse> = {
  encode(message: LiveChatMessageListResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.offlineAt !== undefined && message.offlineAt !== "") {
      writer.uint32(18).string(message.offlineAt);
    }
    if (message.nextPageToken !== undefined && message.nextPageToken !== "") {
      writer.uint32(804818).string(message.nextPageToken);
    }
    for (const v of message.items) {
      LiveChatMessage.encode(v!, writer.uint32(8058).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessageListResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessageListResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.offlineAt = reader.string();
          continue;
        }
        case 100602: {
          if (tag !== 804818) {
            break;
          }

          message.nextPageToken = reader.string();
          continue;
        }
        case 1007: {
          if (tag !== 8058) {
            break;
          }

          message.items.push(LiveChatMessage.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessageListResponse {
    return {
      offlineAt: isSet(object.offlineAt)
        ? globalThis.String(object.offlineAt)
        : isSet(object.offline_at)
        ? globalThis.String(object.offline_at)
        : "",
      nextPageToken: isSet(object.nextPageToken)
        ? globalThis.String(object.nextPageToken)
        : isSet(object.next_page_token)
        ? globalThis.String(object.next_page_token)
        : "",
      items: globalThis.Array.isArray(object?.items) ? object.items.map((e: any) => LiveChatMessage.fromJSON(e)) : [],
    };
  },

  toJSON(message: LiveChatMessageListResponse): unknown {
    const obj: any = {};
    if (message.offlineAt !== undefined && message.offlineAt !== "") {
      obj.offlineAt = message.offlineAt;
    }
    if (message.nextPageToken !== undefined && message.nextPageToken !== "") {
      obj.nextPageToken = message.nextPageToken;
    }
    if (message.items?.length) {
      obj.items = message.items.map((e) => LiveChatMessage.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessageListResponse>, I>>(base?: I): LiveChatMessageListResponse {
    return LiveChatMessageListResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessageListResponse>, I>>(object: I): LiveChatMessageListResponse {
    const message = createBaseLiveChatMessageListResponse();
    message.offlineAt = object.offlineAt ?? "";
    message.nextPageToken = object.nextPageToken ?? "";
    message.items = object.items?.map((e) => LiveChatMessage.fromPartial(e)) || [];
    return message;
  },
};

function createBaseLiveChatMessage(): LiveChatMessage {
  return { id: "", snippet: undefined, authorDetails: undefined };
}

export const LiveChatMessage: MessageFns<LiveChatMessage> = {
  encode(message: LiveChatMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(810).string(message.id);
    }
    if (message.snippet !== undefined) {
      LiveChatMessageSnippet.encode(message.snippet, writer.uint32(18).fork()).join();
    }
    if (message.authorDetails !== undefined) {
      LiveChatMessageAuthorDetails.encode(message.authorDetails, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LiveChatMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLiveChatMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 101: {
          if (tag !== 810) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.snippet = LiveChatMessageSnippet.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.authorDetails = LiveChatMessageAuthorDetails.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LiveChatMessage {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      snippet: isSet(object.snippet) ? LiveChatMessageSnippet.fromJSON(object.snippet) : undefined,
      authorDetails: isSet(object.authorDetails)
        ? LiveChatMessageAuthorDetails.fromJSON(object.authorDetails)
        : isSet(object.author_details)
        ? LiveChatMessageAuthorDetails.fromJSON(object.author_details)
        : undefined,
    };
  },

  toJSON(message: LiveChatMessage): unknown {
    const obj: any = {};
    if (message.id !== undefined && message.id !== "") {
      obj.id = message.id;
    }
    if (message.snippet !== undefined) {
      obj.snippet = LiveChatMessageSnippet.toJSON(message.snippet);
    }
    if (message.authorDetails !== undefined) {
      obj.authorDetails = LiveChatMessageAuthorDetails.toJSON(message.authorDetails);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LiveChatMessage>, I>>(base?: I): LiveChatMessage {
    return LiveChatMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LiveChatMessage>, I>>(object: I): LiveChatMessage {
    const message = createBaseLiveChatMessage();
    message.id = object.id ?? "";
    message.snippet = (object.snippet !== undefined && object.snippet !== null)
      ? LiveChatMessageSnippet.fromPartial(object.snippet)
      : undefined;
    message.authorDetails = (object.authorDetails !== undefined && object.authorDetails !== null)
      ? LiveChatMessageAuthorDetails.fromPartial(object.authorDetails)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}